fix: Consolidate MultiSizeAddr trait usage across EEPROM modules
