+++ src/storage_async.rs
@@ -1,4 +1,4 @@
-use crate::{eeprom24x::MultiSizeAddr, eeprom24x_async::AsyncPageWrite, Eeprom24x, Error, Storage};
+use crate::{MultiSizeAddr, eeprom24x_async::AsyncPageWrite, Eeprom24x, Error, Storage};
 use core::cmp::min;
 
 use embedded_hal_async::{delay::DelayNs as AsyncDelayNs, i2c::I2c as AsyncI2c};
